feat: Seperate `views` logic
- Seperated constant and functions to external files
- Moved the flask admin to another file
